[SOK-36] refactor: перенес движения игрока в gameloop
