+++ packages/client/src/components/Game/gameLoop.tsx
@@ -1,23 +1,31 @@
 import { HandlePlayerHit, resetPlayerPosition } from './player'
 import { updateEnemyPositions, respawnEnemies } from './enemy'
 import { clearCanvas, drawPlayer, drawEnemies, drawObstacles } from './utils'
-import { Enemy, Obstacle, Player } from '@/components/Game/gameTypes'
+import {
+  ControlsProps,
+  Enemy,
+  Obstacle,
+  Player,
+} from '@/components/Game/gameTypes'
 import { detectEnemyCollision } from '@/components/Game/collision'
+import { updatePlayerMovement } from '@/components/Game/controls'
 
 /**
  * Основной игровой цикл, который обновляет состояние игры и перерисовывает экран каждый кадр.
  * @param context - Контекст рисования для Canvas.
+ * @param canvasRef - Ссылка на Canvas.
  * @param playerRef - Ссылка на текущего игрока.
  * @param enemiesRef - Ссылка на массив врагов.
- * @param obstacles - Массив препятствий.
+ * @param obstaclesRef - Ссылка на массив препятствий.
  * @param livesRef - Ссылка на текущее количество жизней игрока.
  * @param handleGameOver - Обработчик события окончания игры.
  */
 export const gameLoop = (
   context: CanvasRenderingContext2D,
+  canvasRef: React.MutableRefObject<HTMLCanvasElement | null>,
   playerRef: React.MutableRefObject<Player>,
   enemiesRef: React.MutableRefObject<Enemy[]>,
-  obstacles: Obstacle[],
+  obstaclesRef: React.MutableRefObject<Obstacle[]>,
   livesRef: React.MutableRefObject<number>,
   handleGameOver: () => void
 ) => {
@@ -25,9 +33,16 @@ export const gameLoop = (
 
   // Обновление позиций врагов
   updateEnemyPositions(playerRef.current, enemiesRef)
-
+  if (!canvasRef.current) return
+  const moveProps: ControlsProps = {
+    playerRef,
+    obstacles: obstaclesRef.current,
+    canvasWidth: canvasRef.current.width,
+    canvasHeight: canvasRef.current.height,
+  }
+  updatePlayerMovement(moveProps)
   // Отрисовка всех игровых объектов
-  drawObstacles(context, obstacles)
+  drawObstacles(context, obstaclesRef.current)
   drawPlayer(context, playerRef.current)
   drawEnemies(context, enemiesRef.current)
